Refactor utils and config
Extract a utils.go and config.go to declutter the original pastad.go
file.
